msm7627a: bluetooth: handle BT On failure case proprely
If Bluetooth on is failed due to some reason, clean up is not
proper.
Also when BT turn on fail, we are trying to release some resources
twice. Once as a part of handling the fail cleanup and other
when the userspace initiates turn off process because of turn
on timeout.

CRs-fixed: 439129
Change-Id: I5a5bd35e8f3ba9cfebfd8f2b18b517180df95d39
Signed-off-by: Ram Mohan Korukonda <rkorukon@codeaurora.org>